+++ changelog.md
@@ -5,6 +5,7 @@ Version 18.08
 ===
 
 + Add "same link" operator for "VPoint" class.
++ Fix bug of BFGS transform in tinycadlib.
 
 Version 18.07
 ===